WIP commit - ensuring that resolver response is cached
I'm making sure that the resolver stores the API response somewhere for the
selector to pick it up, to ensure that we get posts when re-doing a previous
query. Next step is going to be to make sure all blocks get notified of new post
IDs to exclude so they re-render too. Then a massive cleanup.