feat(homepage-posts): add a filter to enable duplicating posts (#886)
since we have a deduplication process for posts, we can't show the same post in two different
homepage-posts blocks. Which should be good for all cases. Using the newspack-popups and setting a
popup with a homepage-posts block will show the non-duplicated posts in the homepage and the popup
in the first time we're loading the page, but the second time (we won't have the popup since it's
showed only once or once a day) we'll have missing (robbed) posts from the homepage-posts block of
the page as explained here https://github.com/Automattic/newspack-popups/issues/126. We need to add
an exception for the deduplication system for such case.