fix(author-profile): only display guest author profile when linked (#1688)
This PR addresses an issue where it was possible to select a WP user account that was linked to a guest author account in the author profile block. Previously we weren't correctly filtering out linked WP users in the block whenever a search term was input that matched the WP user but not the linked guest author.