fix(ras-acc): only trigger update_checkout event on validation (#1823)
This PR resolves an issue where the update_checkout woo event causes the order review table to end up in a blocked state. This looks to be happening because we are hiding the original order review table, cloning it, and rendering it in a non-standard part of the checkout. This seems to be happening at the same time as Woo attempts to block the table, perform some Woo logic, and then finally unblock the table and as a result the cloned table ends up in a blocked state.

We fix this by removing the call to trigger update_checkout event from the setEditingDetails method as well as from the cover fees script in Automattic/newspack-plugin#3330 to allow updating checkout to be managed entirely within the validateForm function.