fix(ras-acc): use config flag to skip auth newsletters signup (#1874)
This PR addresses an issue where if checkout data is present in local storage, the auth flow will always skip the newsletters signup step. We fix this by adding an explicit skipNewslettersSignup flag to the auth modal config object and using this to decide whether to bypass the newsletters modal instead.