fix: account for empty space in error messages (#1959)
This PR resolves an issue where billing field errors were not rendering below the billing field inputs as expected. This is because our error handler first checks for a non HTML error, relying on the first character of the error message being <. However, in some cases, the HTML string Woo returns begins with a space.

We resolve this by stripping beginning white space when checking the first character of the error message.